backend/requirements.txt
Flask==1.1.1
stripe==2.55.1
PyJWT==2.0.0
pymongo[srv]==3.11.2
python-dotenv==0.15.0                              
gunicorn==20.0.4
Flask-Cors==3.0.10